exercicio 37 mudificado nome da classe , criada varia ve char codigo refatorado
